Conflicts:
	server/server_config.json

+++ aws_provisioning/create_mr_node.py
@@ -15,13 +15,6 @@ from fabric.exceptions import NetworkError
 import sys
 import aws_util
 
-
-def get_running_instance(config):
-    conn = aws_util.connect_cfg(config)
-    instance = aws_util.get_instance(conn, config["instance_id"])
-    print "Instance", instance.id, "is", instance.state
-    return conn, instance
-
 def bootstrap_instance(config, instance):
     ssl_user = config.get("ssl_user", "ubuntu")
     ssl_key_path = config.get("ssl_key_path", "~/.ssh/id_rsa.pub")
@@ -36,6 +29,11 @@ def bootstrap_instance(config, instance):
     #sudo("apt-get --yes dist-upgrade")
     sudo('pip install simplejson scales boto')
 
+    base_dir = config.get("base_dir", "/mnt/telemetry")
+    # Put work dirs in /mnt where there's plenty of space:
+    sudo("mkdir -p " + base_dir)
+    sudo("chown %s:%s %s" % (ssl_user, ssl_user, base_dir))
+
     home = "/home/" + ssl_user
     print "Preparing MR code"
     with cd(home):
@@ -43,7 +41,7 @@ def bootstrap_instance(config, instance):
         run("git clone https://github.com/sstoiana/s3funnel.git")
     with cd(home + "/s3funnel"):
         sudo("python setup.py install")
-    with cd(home + "/telemetry-server"):
+    with cd(base_dir):
         # "data" is a dummy dir just to give it somewhere to look for local data.
         run("mkdir job work data")
 
@@ -51,19 +49,24 @@ def run_mapreduce(config, instance):
     ssl_user = config.get("ssl_user", "ubuntu")
     home = "/home/" + ssl_user
     mr_cfg = config["mapreduce"]
+    base_dir = config.get("base_dir", "/mnt/telemetry")
+    job_dir = base_dir + "/job"
+    data_dir = base_dir + "/data"
+    work_dir = base_dir + "/work"
     with cd(home + "/telemetry-server"):
         job_script = mr_cfg["job_script"]
         input_filter = mr_cfg["input_filter"]
-        put(job_script, "job")
-        put(input_filter, "job")
-        job_script_base = os.path.basename(job_script)
-        input_filter_base = os.path.basename(input_filter)
-        job_args = (job_script_base, input_filter_base, config["aws_key"], config["aws_secret_key"], mr_cfg["data_bucket"])
-        run('python job.py job/%s --input-filter job/%s --data-dir ./data --work-dir ./work --aws-key "%s" --aws-secret-key "%s" --bucket "%s" --output job/output.txt' % job_args)
+        put(job_script, job_dir)
+        put(input_filter, job_dir)
+        job_script_path = "/".join((job_dir, os.path.basename(job_script)))
+        input_filter_path = "/".join((job_dir, os.path.basename(input_filter)))
+        output_path = "/".join((job_dir, "output.txt"))
+        job_args = (job_script_path, input_filter_path, data_dir, work_dir, output_path, config["aws_key"], config["aws_secret_key"], mr_cfg["data_bucket"])
+        run('python job.py %s --input-filter %s --data-dir %s --work-dir %s --output %s --aws-key "%s" --aws-secret-key "%s" --bucket "%s"' % job_args)
         # TODO: consult "output_compression"
-        run("lzma job/output.txt")
+        run("lzma " + output_path)
         # TODO: upload job/output.txt.lzma to S3 output_bucket.output_filename
-        result = get("job/output.txt.lzma", mr_cfg["output_filename"])
+        result = get(output_path + ".lzma", mr_cfg["output_filename"])
         # TODO: check result.succeeded before bailing.
 
 if len(sys.argv) < 2:
@@ -76,10 +79,13 @@ config_file.close()
 
 print "Using the following config:"
 print json.dumps(config)
-#sys.exit(-1)
 
-conn, instance = aws_util.create_instance(config)
-#conn, instance = get_running_instance(config)
+if "instance_id" in config:
+    conn = aws_util.connect_cfg(config)
+    instance = aws_util.get_instance(conn, config["instance_id"])
+else:
+    conn, instance = aws_util.create_instance(config)
+
 try:
     ssl_user = config.get("ssl_user", "ubuntu")
     ssl_key_path = config.get("ssl_key_path", "~/.ssh/id_rsa.pub")
@@ -93,6 +99,7 @@ try:
 
     # Can't connect when using known hosts :(
     env.disable_known_hosts = True
+    env.keepalive = 5
 
     if aws_util.wait_for_ssh(config.get("ssl_retries", 3)):
         print "SSH Connection is ready."
@@ -100,9 +107,16 @@ try:
         print "Failed to establish SSH Connection to", instance.id
         sys.exit(2)
 
-    bootstrap_instance(config, instance)
+    if not config.get("skip_bootstrap", False):
+        bootstrap_instance(config, instance)
+
     run_mapreduce(config, instance)
 finally:
     # All done: Terminate this mofo
-    print "Terminating", instance.id
-    conn.terminate_instances(instance_ids=[instance.id])
+    if "instance_id" in config:
+        # It was an already-existing instance, leave it alone
+        print "Not terminating instance", instance.id
+    else:
+        # we created it ourselves, terminate it.
+        print "Terminating", instance.id
+        conn.terminate_instances(instance_ids=[instance.id])
